Рабочие методы, но выдает ошибку 405

+++ crud-web-app/src/main/java/com/example/repository/AuthorRepository.java
@@ -3,5 +3,5 @@ package com.example.repository;
 
 import com.example.model.Author;
 
-public interface AuthorRepository<K, T> extends DefaultCrudRepository<Author, String, Integer>{
+public interface AuthorRepository<K, T> extends DefaultCrudRepository<Author, String, Long>{
 }
