+++ crud-web-app/src/main/java/com/example/repository/GenreRepository.java
@@ -3,5 +3,5 @@ package com.example.repository;
 
 import com.example.model.Genre;
 
-public interface GenreRepository<T, N extends Number> extends DefaultCrudRepository<Genre, String, Integer>{
+public interface GenreRepository<T, N extends Number> extends DefaultCrudRepository<Genre, String, Long>{
 }
