+++ crud-web-app/src/main/java/com/example/repository/impl/AuthorRepositoryImpl.java
@@ -28,7 +28,7 @@ public class AuthorRepositoryImpl implements AuthorRepository<Author, String> {
     }
 
     @Override
-    public Author findById(Integer id) {
+    public Author findById(Long id) {
         return (Author) entityManager.createNativeQuery("FROM book_list.author where id = \"" + id + "\";", Author.class).getSingleResult();
     }
 
@@ -38,7 +38,7 @@ public class AuthorRepositoryImpl implements AuthorRepository<Author, String> {
     }
 
     @Override
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         entityManager.remove(id);
     }
 }
